+++ src/exec/libf/libft.h
@@ -6,7 +6,7 @@
 /*   By: yabidi <yabidi@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/17 18:16:47 by yabidi            #+#    #+#             */
-/*   Updated: 2022/10/22 11:45:45 by yabidi           ###   ########.fr       */
+/*   Updated: 2023/02/16 19:48:11 by yabidi           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ typedef struct s_list
 	struct s_list	*next;
 }					t_list;
 
-int		ft_atoi(const char *str);
+unsigned long long		ft_atoi(const char *str);
 void	ft_bzero(void *s, size_t n);
 int		ft_isalnum(char t);
 int		ft_isalpha(char t);
